Split the registry + View-as-Text read out of GhosttySurfaceView.swift
workflow-guard-tests file-length fix: move the surface registry
(WeakGhosttySurfaceViewBox, register/unregister/view(for:)) and the
registry-scoped copyableTerminalText read into
GhosttySurfaceRegistry.swift. GhosttySurfaceView.swift drops below its
pre-branch budget (3665 < 3694). outputQueue and the registry become
internal so the same-module registry file keeps the identical
queue/lifetime discipline.

Budget refresh accepts two small spillovers as known debt:
GhosttyRuntime.swift +8 (scrollback-limit config comment) and
WorkspaceDetailView.swift tracked at 512 (View as Text menu + sheet
wiring, cohesive with the view).
